refactor: extract torch adapter and scheduler
